requirements.txt
requests
beautifulsoup4
python-dotenv
google-search-results
sendgrid
ttkbootstrap
pillow
botasaurus
ddgs
selenium
email-validator
